debug

+++ python/FRACluster.py
@@ -255,12 +255,13 @@ def FRACluster(ARGVS, WD, MAX_ITERATION, SUBSAMPLE_SIZE, NODESDIR, THRESHOLD, TH
                         )
                 print("before", fpath2seqcount)
             else:
-                os.mkdir(example_infile_fpath+".split")
-                for infile_path in infile_pathlist:
-                    fpath2seqcount[example_infile_fpath+".split/"+infile_path.split("/")[-1]] = fpath2seqcount[infile_path]
-                    fpath2seqcount.pop(infile_path)
-                    shutil.move(infile_path, example_infile_fpath+".split")
-                splitted_dirpath = example_infile_fpath+".split"
+                if (not os.path.exists(example_infile_fpath+".split")):
+                    os.mkdir(example_infile_fpath+".split")
+                    for infile_path in infile_pathlist:
+                        fpath2seqcount[example_infile_fpath+".split/"+infile_path.split("/")[-1]] = fpath2seqcount[infile_path]
+                        fpath2seqcount.pop(infile_path)
+                        shutil.move(infile_path, example_infile_fpath+".split")
+                    splitted_dirpath = example_infile_fpath+".split"
 
             # rename
             for j, splitted_fname in enumerate(os.listdir(splitted_dirpath)):
